Nuevo function para input de boolean

+++ generator/builder.py
@@ -9,29 +9,32 @@ from generator.uploader.upload import Uploader
 import time
 import optparse
 
-from generator.utils import read_hashes, save_hash
+from generator.utils import bool_from_input, read_hashes, remove_hash, save_hash
 
 
 def both():
     # Pregunta para su vaina
     public = input("Que es tu llave publico: ")
     private = input("Que es tu private key: ")
-    is_this_test_net = input("Prueba? (y/n): ")
-    if is_this_test_net == "y" or is_this_test_net == "Y":
-        is_this_test_net = True
-    else:
-        is_this_test_net = False
+    is_this_test_net = bool_from_input("Prueba? (y/n): ")
     amount = int(input("Cuanto quieres mint? "))
 
     # Abrimos el minter
     minter = Minter(public, private, is_this_test_net)
     start = time.time()
 
+    pre = None
+
     # Loop con amount
     for x in range(amount):
-        print(f"Generating --- {x} de {amount}")
+        print(f"Generating --- {x + 1} de {amount}")
         # Buscamos el id del cock por smart contract
         cock_id = minter.most_recent()
+        
+        while cock_id == pre:
+            cock_id = minter.most_recent()
+        
+        pre = cock_id
         print(f"ID de cock {cock_id}")
         gen = ImageGen(ChickenType.DETAILED_COCK, cock_id, randomifyattributes(Attribute.GEN_0))
         mck = gen.draw()
@@ -45,49 +48,56 @@ def both():
         # Chequeamos el upload no funciono
         if not h:
             # Di lo y termina!
-            print(f"Una problema con {x}. Terminando...")
+            print(f"Una problema con {x + 1}. Terminando...")
             break
         # Guarda hash para futuro
         save_hash(h)
         # Verifica el mint del smart contract con el id y nombre!!!
-        pos = input(f"Hacemos un mint con {mck} quien tiene id de {cock_id} en el smart contract? (y/n)")
-        if pos == "Y" or pos == "y":
+        pos = bool_from_input(f"Hacemos un mint con {mck} quien tiene id de {cock_id} en el smart contract? (y/n) ")
+        if pos:
             res = minter.mint(h)
             # Chequea si no funciono
             if not res:
                 # Di al usario y termina!!
-                print(f"No podiamos hacer mint con {h} con {x}.\nTerminando...")
+                print(f"No podiamos hacer mint con {h} con {x + 1}.\nTerminando...")
                 break
         else:
             print("Ok... Terminando...")
+            # Quita el hash
+            uploader.unpin(h)
+            remove_hash(h)
             break
     
     end = time.time()
     print(f"Tiempo tomado {int(end - start)} segundos")
 
-def generator(amount):
+def generator(amount, save):
+    if save:
+        print("Vamos a guardar!")
+    
     start = time.time()
     for x in range(amount):
-        print(f"Generating --- {x} de {amount}")
+        print(f"Generating --- {x + 1} de {amount}")
         # Abre el class de generation
         gen = ImageGen(ChickenType.DETAILED_COCK, x, randomifyattributes(Attribute.GEN_0))
         mck = gen.draw()
         
-        # Uploaderlo
-        uploader = Uploader(
-            gen.chicken_type,
-            mck,
-            AttributesBuilder.pretty_attributes(gen.color_data, gen.attributes)
-        )
-        # Busca hash
-        _hash = uploader.upload()
-        # Chequea que hash es falso
-        if not _hash:
-            print(f"Una problema con {mck} borrando...")
-            os.unlink(mck + ".png")
-            continue
-        # Guarda hash
-        save_hash(_hash)
+        if save:
+            # Uploaderlo
+            uploader = Uploader(
+                gen.chicken_type,
+                mck,
+                AttributesBuilder.pretty_attributes(gen.color_data, gen.attributes)
+            )
+            # Busca hash
+            _hash = uploader.upload()
+            # Chequea que hash es falso
+            if not _hash:
+                print(f"Una problema con {mck} borrando...")
+                os.unlink(mck + ".png")
+                continue
+            # Guarda hash
+            save_hash(_hash)
     # Busca cuando se termino
     end = time.time()
     print(f"Tiempo tomado para {amount} fue {int(end - start)} segundos")
@@ -133,7 +143,8 @@ def main():
     
     if method.lower() == "generate":
         amount = input("Cuantos vamos a generar? ")
-        generator(int(amount))
+        save = bool_from_input("Vamos a guardar? ")
+        generator(int(amount), save)
     elif method.lower() == "mint":
         minter()
     elif method.lower() == "upload":
